requirements.txt
asgiref==3.2.3
astroid==2.3.3
automium==0.2.6
automium-web==0.1.1
autopep8==1.4.4
bcrypt==3.1.7
cffi==1.13.2
Click==7.0
Django==3.0.2
django-appconf==1.0.3
django-compressor==2.4
django-filter==2.2.0
django-libsass==0.8
django-model-utils==4.0.0
djangorestframework==3.11.0
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
legacy==0.1.6
libsass==0.15.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
passlib==1.7.2
pycodestyle==2.5.0
pycparser==2.19
pylint==2.4.4
python-decouple==3.3
pytz==2019.3
rcssmin==1.0.6
reverse==0.1.0
rjsmin==1.1.0
six==1.13.0
sqlparse==0.3.0
Werkzeug==0.16.0
wrapt==1.11.2
